tests/wpt/tests/compat/webkit-mask-box-enumeration.html

<!DOCTYPE html>
<title>WebKitMaxBoxImage on CSSStyleDeclaration</title>
<link rel="help" href="https://compat.spec.whatwg.org/#css-simple-aliases">
<meta name="assert" content="This test verifies that these WebKit properties are enumerated on CSSStyleDeclaration" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
</body>
<script>
const PREFIXED_PROPS = [
  'webkitMaskBoxImage',
  'webkitMaskBoxImageOutset',
  'webkitMaskBoxImageRepeat',
  'webkitMaskBoxImageSlice',
  'webkitMaskBoxImageSource',
  'webkitMaskBoxImageWidth',
];

const docBodyStyle = document.body.style;
for (let prop of PREFIXED_PROPS) {
  test(() => {
    assert_true(prop in docBodyStyle);
  }, `${prop} found on CSSStyleDeclaration`);
}
</script>
